feat: updates example with clearer naming, groups example, descriptions and typo fixes
