+++ examples/advanced-project/checkly.config.ts
@@ -1,21 +1,42 @@
 import { defineConfig } from '@checkly/cli'
 
+/**
+ * See https://www.checklyhq.com/docs/cli/project-structure/
+ */
 const config = defineConfig({
+  /* A human friendly name for your project */
   projectName: 'Advanced Example Project',
+  /** A logical ID that needs to be unique across your Checkly account,
+   * See https://www.checklyhq.com/docs/cli/constructs/ to learn more about logical IDs.
+   */
   logicalId: 'advanced-example-project',
+  /* An optional URL to your Git repo */
   repoUrl: 'https://github.com/checkly/checkly-cli',
+  /* Sets default values for Checks */
   checks: {
+    /* Checkly datacenters where to run your Checks as monitors */
     locations: ['us-east-1', 'eu-west-1'],
+    /* An optional array of tags to organize your Checks */
     tags: ['mac'],
+    /** The Checkly Runtime identifier, determining NPM packages and the Node.js version available at runtime.
+    * See https://www.checklyhq.com/docs/cli/npm-packages/
+    */
     runtimeId: '2022.10',
-    checkMatch: '**/*.check.ts',
+    /* A glob pattern that matches the Checks inside your repo, see https://www.checklyhq.com/docs/cli/using-check-test-match/ */
+    checkMatch: '**/__checks__/**/*.check.ts',
     browserChecks: {
-      testMatch: '**/__checks__/*.spec.ts', // this matches any Playwright spec-files and automagically creates a Browser check
+      /* A glob pattern matches any Playwright .spec.ts files and automagically creates a Browser Check. This way, you
+      * can just write native Playwright code. See https://www.checklyhq.com/docs/cli/using-check-test-match/
+      * */
+      testMatch: '**/__checks__/**/*.spec.ts',
     },
   },
   cli: {
+    /* The default datacenter location to use when running npx checkly test */
     runLocation: 'eu-west-1',
   },
 })
+
+
 
 export default config
