+++ examples/advanced-project/package.json
@@ -13,6 +13,7 @@
   "license": "ISC",
   "devDependencies": {
     "@checkly/cli": "latest",
+    "@playwright/test": "^1.28.0",
     "ts-node": "10.9.1",
     "typescript": "4.9.5"
   }
