+++ examples/advanced-project/src/__checks__/api.check.ts
@@ -1,0 +1,16 @@
+import { ApiCheck } from '@checkly/cli/dist/constructs'
+import { websiteGroup } from './website.check-group'
+
+new ApiCheck('check-group-api-check-1', {
+  name: 'Fetch stats for homepage',
+  group: websiteGroup,
+  degradedResponseTime: 10000,
+  maxResponseTime: 20000,
+  request: {
+    method: 'GET',
+    url: 'https://api.checklyhq.com/public-stats',
+    followRedirects: true,
+    skipSSL: false,
+    assertions: []
+  }
+})
