+++ examples/advanced-project/src/services/api/__checks__/api.check.ts
@@ -1,8 +1,12 @@
 import { ApiCheck, AssertionBuilder } from '@checkly/cli/constructs'
-import { slackChannel, webhookChannel } from '../../alert-channels'
+import { slackChannel, webhookChannel } from '../../../alert-channels'
 
-new ApiCheck('homepage-api-check-1', {
-  name: 'Homepage - fetch stats',
+/**
+ * This API Check is picked up using the defaults from the checkly.config.ts file at the root of this example.
+ */
+
+new ApiCheck('example-api-check-not-grouped-1', {
+  name: 'API Check not in a Group',
   alertChannels: [slackChannel, webhookChannel],
   degradedResponseTime: 10000,
   maxResponseTime: 20000,
